Develop

+++ appshop/api/views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.models import User
+from django.core.mail import send_mail
 from django.db.models import Q
 from django.http import HttpResponseRedirect
 from django.urls import reverse
@@ -9,10 +10,8 @@ from rest_framework.response import Response
 from rest_framework.generics import GenericAPIView
 from rest_framework.pagination import PageNumberPagination
 from django.shortcuts import get_object_or_404
-
 from appshop.api.serializers import DetailItemSerializer, ShortDetailItemSerializer, ModifyDetailItemSerializer
 from appshop.models import Item
-from appshop.views import get_landing_page
 
 
 class ItemPaginator(PageNumberPagination):
@@ -25,7 +24,7 @@ class ItemListUserAPI(GenericAPIView):
 
     def get(self, request, token):
         user = Token.objects.get(key=token).user
-        items = Item.objects.filter(seller=user.username) | Item.objects.filter(buyer=user.username)
+        items = Item.objects.filter(seller=user) | Item.objects.filter(buyer=user)
         items = items.distinct().order_by("-created_date")
 
         serializer = DetailItemSerializer(items, many=True)
@@ -38,10 +37,13 @@ class ItemListUserAPI(GenericAPIView):
         serializer.is_valid(raise_exception=True)
         data = serializer.validated_data
 
-        item = Item(title=data["title"], description=data["description"], price=data["price"],
-                    seller=user.username, buyer='', status='WAITING')
-        item.save()
-        return Response({"message": "item created"})
+        if data['price'] >= 0 :
+            item = Item(title=data["title"], description=data["description"], price=data["price"],
+                        seller=user, status='WAITING')
+            item.save()
+            return Response({"message": "item created"})
+        else:
+            return Response({"message": "negative price is not allowed"}, 500)
 
 
 class ItemListSaleUserAPI(GenericAPIView):
@@ -51,7 +53,7 @@ class ItemListSaleUserAPI(GenericAPIView):
 
     def get(self, request, token):
         user = Token.objects.get(key=token).user
-        items = Item.objects.filter(Q(status='SALE'), ~Q(seller=user.username))
+        items = Item.objects.filter(Q(status='SALE'), ~Q(seller=user))
         items = items.order_by("-created_date")
 
         page = self.paginate_queryset(items)
@@ -70,7 +72,7 @@ class ItemListSaleUserSearchAPI(GenericAPIView):
 
     def get(self, request, token, search):
         user = Token.objects.get(key=token).user
-        items = Item.objects.filter(Q(status='SALE'), ~Q(seller=user.username))
+        items = Item.objects.filter(Q(status='SALE'), ~Q(seller=user))
         items = items.filter(title__contains=search)
         items = items.order_by("-created_date")
 
@@ -120,6 +122,19 @@ class ItemListSearchAPI(GenericAPIView):
             return self.get_paginated_response([])
 
 
+def notify_email(seller, to, title, price):
+    subject = 'New item sold !' if seller else 'New item bought !'
+
+    message = 'You ' + ('sold' if seller else 'bought') + ' the item ' + title + ' (' + str(price) + '€).'
+    message = message + ' See you soon in our Webshop !'
+    send_mail(
+        subject,
+        message,
+        'no-reply@webshop.com',
+        [to],
+    )
+
+
 class ItemDetailsAPI(GenericAPIView):
     pagination_class = ItemPaginator
     permission_classes = [IsAuthenticatedOrReadOnly]
@@ -133,29 +148,57 @@ class ItemDetailsAPI(GenericAPIView):
     def put(self, request, item_id):
         item = get_object_or_404(Item, pk=item_id)
 
+        token = request.headers['Authorization'][6:]
+
         serializer = ModifyDetailItemSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         data = serializer.validated_data
 
-        item.price = data["price"]
-        item.title = data["title"]
-        item.description = data["description"]
-        item.status = data["status"]
-        item.save()
+        if item.status == 'SOLD':
+            return Response({"message": "You can not modify an already sold item"})
 
-        return Response({"message": "item updated"})
+        if item.seller != Token.objects.get(key=token).user and data["status"] != 'SOLD':
+            return Response({"message": "You do not own this item"})
+
+        else:
+            if item.seller == Token.objects.get(key=token).user and data["status"] == 'SOLD':
+                return Response({"message": "You can not buy your own item"})
+            else:
+                if(data["price"]<0):
+                    return Response({"message": "negative price is not allowed"},500)
+
+                item.price = data["price"]
+                item.title = data["title"]
+                item.description = data["description"]
+                item.status = data["status"]
+
+                # buy
+                if data["status"] == 'SOLD':
+                    user = Token.objects.get(key=token).user
+                    item.buyer = user
+                    notify_email(False, user.email, item.title, item.price)
+                    notify_email(True, item.seller.email, item.title, item.price)
+
+                item.save()
+
+                return Response({"message": "item updated"})
 
     def delete(self, request, item_id):
         item = get_object_or_404(Item, pk=item_id)
-        item.delete()
-        return Response({"message": "item deleted"})
+        token = request.headers['Authorization'][6:]
+
+        if (item.status == 'SOLD'):
+            return Response({"message": "You can not delete an already sold item"})
+
+        if (item.seller != Token.objects.get(key=token).user):
+            return Response({"message": "You do not own this item"})
+
+        else:
+            item.delete()
+            return Response({"message": "item deleted"})
 
 
 class PopulateAPI(GenericAPIView):
-
-    def get(self, request):
-        return get_landing_page(request)
-
     def post(self, request):
         Item.objects.all().delete()
         User.objects.all().delete()
@@ -170,49 +213,49 @@ class PopulateAPI(GenericAPIView):
         for u in [user1, user2, user3, user4, user5, user6]:
             Token.objects.create(user=u)
 
-        item0 = Item(title="piano", description="with tiles (black and white)!", price=500, seller="testuser1",
-                     status='WAITING')
-        item1 = Item(title="drums", description="very good quality", price=60, seller="testuser1", status='SALE')
-        item2 = Item(title="table", description="with few scratches", price=20, seller="testuser1", status='SALE')
-        item3 = Item(title="chair", description="very comfortable", price=40, seller="testuser1", status='SALE')
-        item4 = Item(title="desk", description="few pen traces", price=30, seller="testuser1", status='SALE')
-        item5 = Item(title="kitchen stuff", description="used only during 3 months", price=10, seller="testuser1",
+        item0 = Item(title="piano", description="with tiles (black and white)!", price=500, seller=user1,
                      status='SALE')
-        item6 = Item(title="bed", description="almost new", price=400, seller="testuser1", status='SALE')
-        item7 = Item(title="curtains", description="with nice drawings", price=20, seller="testuser1", status='SALE')
-        item8 = Item(title="car", description="Renault one", price=900, seller="testuser1", status='SALE')
-        item9 = Item(title="lamp", description="nice lamp to read", price=10, seller="testuser1", status='SALE')
+        item1 = Item(title="drums", description="very good quality", price=60, seller=user1, status='SALE')
+        item2 = Item(title="table", description="with few scratches", price=20, seller=user1, status='SALE')
+        item3 = Item(title="chair", description="very comfortable", price=40, seller=user1, status='SALE')
+        item4 = Item(title="desk", description="few pen traces", price=30, seller=user1, status='SALE')
+        item5 = Item(title="kitchen stuff", description="used only during 3 months", price=10, seller=user1,
+                     status='SALE')
+        item6 = Item(title="bed", description="almost new", price=400, seller=user1, status='SALE')
+        item7 = Item(title="curtains", description="with nice drawings", price=20, seller=user1, status='SALE')
+        item8 = Item(title="car", description="Renault one", price=900, seller=user1, status='SALE')
+        item9 = Item(title="lamp", description="nice lamp to read", price=10, seller=user1, status='SALE')
 
-        item10 = Item(title="piano", description="with tiles (black and white)!", price=499, seller="testuser2",
-                      status='WAITING')
-        item11 = Item(title="drums", description="very good quality", price=59, seller="testuser2", status='SALE')
-        item12 = Item(title="table", description="with few scratches", price=19, seller="testuser2", status='SALE')
-        item13 = Item(title="chair", description="very comfortable", price=39, seller="testuser2", status='SALE')
-        item14 = Item(title="desk", description="few pen traces", price=29, seller="testuser2", status='SALE')
-        item15 = Item(title="kitchen stuff", description="used only during 3 months", price=9, seller="testuser2",
+        item10 = Item(title="piano", description="with tiles (black and white)!", price=499, seller=user2,
                       status='SALE')
-        item16 = Item(title="bed", description="almost new", price=399, seller="testuser2", status='SALE')
-        item17 = Item(title="curtains", description="with nice drawings", price=19, seller="testuser2", status='SALE')
-        item18 = Item(title="car", description="Renault one", price=899, seller="testuser2", status='SALE')
-        item19 = Item(title="lamp", description="nice lamp to read", price=9, seller="testuser2", status='SALE')
+        item11 = Item(title="drums", description="very good quality", price=59, seller=user2, status='SALE')
+        item12 = Item(title="table", description="with few scratches", price=19, seller=user2, status='SALE')
+        item13 = Item(title="chair", description="very comfortable", price=39, seller=user2, status='SALE')
+        item14 = Item(title="desk", description="few pen traces", price=29, seller=user2, status='SALE')
+        item15 = Item(title="kitchen stuff", description="used only during 3 months", price=9, seller=user2,
+                      status='SALE')
+        item16 = Item(title="bed", description="almost new", price=399, seller=user2, status='SALE')
+        item17 = Item(title="curtains", description="with nice drawings", price=19, seller=user2, status='SALE')
+        item18 = Item(title="car", description="Renault one", price=899, seller=user2, status='SALE')
+        item19 = Item(title="lamp", description="nice lamp to read", price=9, seller=user2, status='SALE')
 
-        item20 = Item(title="piano", description="with tiles (black and white)!", price=505, seller="testuser3",
-                      status='WAITING')
-        item21 = Item(title="drums", description="very good quality", price=65, seller="testuser3", status='SALE')
-        item22 = Item(title="table", description="with few scratches", price=25, seller="testuser3", status='SALE')
-        item23 = Item(title="chair", description="very comfortable", price=45, seller="testuser3", status='SALE')
-        item24 = Item(title="desk", description="few pen traces", price=35, seller="testuser3", status='SALE')
-        item25 = Item(title="kitchen stuff", description="used only during 3 months", price=15, seller="testuser3",
+        item20 = Item(title="piano", description="with tiles (black and white)!", price=505, seller=user3,
                       status='SALE')
-        item26 = Item(title="bed", description="almost new", price=405, seller="testuser3", status='SALE')
-        item27 = Item(title="curtains", description="with nice drawings", price=15, seller="testuser3", status='SALE')
-        item28 = Item(title="car", description="Renault one", price=905, seller="testuser3", status='SALE')
-        item29 = Item(title="lamp", description="nice lamp to read", price=15, seller="testuser3", buyer="testuser1",
-                      status='SOLD')
+        item21 = Item(title="drums", description="very good quality", price=65, seller=user3, status='SALE')
+        item22 = Item(title="table", description="with few scratches", price=25, seller=user3, status='SALE')
+        item23 = Item(title="chair", description="very comfortable", price=45, seller=user3, status='SALE')
+        item24 = Item(title="desk", description="few pen traces", price=35, seller=user3, status='SALE')
+        item25 = Item(title="kitchen stuff", description="used only during 3 months", price=15, seller=user3,
+                      status='SALE')
+        item26 = Item(title="bed", description="almost new", price=405, seller=user3, status='SALE')
+        item27 = Item(title="curtains", description="with nice drawings", price=15, seller=user3, status='SALE')
+        item28 = Item(title="car", description="Renault one", price=905, seller=user3, status='SALE')
+        item29 = Item(title="lamp", description="nice lamp to read", price=15, seller=user3,
+                      status='SALE')
 
         for item in [item0, item1, item2, item3, item4, item5, item6, item7, item8, item9,
                      item10, item11, item12, item13, item14, item15, item16, item17, item18, item19,
                      item20, item21, item22, item23, item24, item25, item26, item27, item28, item29]:
             item.save()
 
-        return HttpResponseRedirect(reverse("home"))  # should redirect to a different URL !!!
+        return HttpResponseRedirect(reverse("home"))
